requirements.txt
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.0.0
gitpython>=3.1.0
click>=8.0.0
pyyaml>=6.0
colorama>=0.4.0
flask>=3.0.0
flask-cors>=6.0.0
requests>=2.32.0